modified:   WebDevelopmentLab03/pages/DJ_Helper.py

+++ WebDevelopmentLab03/pages/DJ_Helper.py
@@ -66,7 +66,7 @@ with st.form("survey_form"):
                         if float(i["artist"]["name"]) != inputArtist:
                             similarSongList.append(i["name"])
                     elif not boolExpressionArtist and boolExpressionPlays:
-                        if float(i["playcount"]) >= 100000.0:
+                        if int(i["playcount"]) >= 100000:
                             similarSongList.append(i["name"])
                     elif boolExpressionArtist and boolExpressionPlays:
                         if float(i["artist"]["name"]) != inputArtist:
